Applying new pattern

+++ src/services/unit_type.service.js
@@ -1,70 +1,69 @@
+const HttpError = require('../classes/HttpError');
 const unitTypesRepository = require('../repositories/unit_types.repository');
 
 const getAll = async () => {
-    const data = await unitTypesRepository.getAll();
-
-    if (data && data.length === 0) {
-        return {
-            message: 'Nenhuma Unidade de Medida foi encontrada!',
-            data: []
-        }
-    }
-
-    return {
-        message: 'Unidades de Medida encontradas com sucesso!',
-        data: data
-    }
+    return unitTypesRepository.getAll();
 }
 
 const getById = async (unitTypeId) => {
     const data = await unitTypesRepository.getById(unitTypeId);
 
-    return {
-        message: 'Unidade de Medida encontrada com sucesso!',
-        data: data
-    }
+    if (!data)
+        throw new HttpError('Unidade de medida inexistente!', 400);
+
+    return data;
 }
 
 const create = async (body) => {
-    const bodyFormatted = formatBody(body);
+    const bodyFormatted = normalizeData(body);
 
-    const data = await unitTypesRepository.create(bodyFormatted);
+    const unitTypeExist = await unitTypesRepository.getByNameOrType(bodyFormatted.name, bodyFormatted.type);
 
-    return {
-        message: 'Unidade de Medida cadastrada com sucesso!',
-        data: data
-    }
+    if (unitTypeExist)
+        throw new HttpError('Já existe uma unidade de medida com este nome ou com esta abreviação', 409);
+
+    const newUnitType = await unitTypesRepository.create(bodyFormatted);
+
+    return unitTypesRepository.getById(newUnitType.id);
 }
 
 const update = async (body, unitTypeId) => {
-    const bodyFormatted = formatBody(body);
+    const unitTypeExist = await unitTypesRepository.getById(unitTypeId);
 
-    const data = await unitTypesRepository.update(bodyFormatted, unitTypeId);
+    if (!unitTypeExist)
+        throw new HttpError('Unidade de medida inexistente!', 400);
 
-    return {
-        message: 'Unidade de Medida atualizada com sucesso!',
-        data: data
-    }
+    const name = body.name ? body.name : '';
+    const type = body.type ? body.type : '';
+
+    const unitTypeExistNameOrType = await unitTypesRepository.getByNameOrType(name, type, unitTypeId);
+
+    if (unitTypeExistNameOrType)
+        throw new HttpError('Já existe uma unidade de medida com este nome ou com esta abreviação', 409);
+
+    const bodyFormatted = normalizeData(body, unitTypeExist);
+
+    await unitTypesRepository.update(bodyFormatted, unitTypeId);
+
+    return unitTypesRepository.getById(unitTypeId);
 }
 
 const remove = async (unitTypeId) => {
-    await unitTypesRepository.remove(unitTypeId);
+    const unitTypeExist = await unitTypesRepository.getById(unitTypeId);
 
-    return {
-        message: 'Unidade de Medida excluída com sucesso!'
-    }
+    if (!unitTypeExist)
+        throw new HttpError('Unidade de medida inexistente!', 400);
+
+    return unitTypesRepository.remove(unitTypeId);
 }
 
-function formatBody (body) {
-    const unitTypeName = body.name.trim();
-    const unitTypeType = body.type.trim();
-
-    const bodyFormatted = {
-        name: unitTypeName,
-        type: unitTypeType
+function normalizeData (body, unitType = null) {
+    return {
+        name: body.name ? body.name.trim() 
+            : unitType?.name ? unitType.name : '',
+        type: body.type ? body.type.trim() 
+            : unitType?.type ? unitType.type : ''
     }
-
-    return bodyFormatted;
 }
 
 module.exports = {
